+++ lib/resource/template/scrollview/ScrollViewPage.dart
@@ -124,7 +124,7 @@ class ScrollViewPage {
     if( thumbVisibility && isNotWeb ) {
       scrollBar = Scrollbar(
         child: scrollChild,
-        isAlwaysShown: true,
+        // isAlwaysShown: true,  //removed from flutter sdk 3.13.1
         trackVisibility: true,
       );
     } else {